requirements.txt
pytest
sqlalchemy
psycopg2-binary
python-dotenv